// frontend/src/api/generated/schemas/strategyResponse.ts
// @ts-nocheck
// This file is auto-generated by orval. Do not edit manually.
import type { StrategyResponseDescription } from "./strategyResponseDescription";
import type { StrategyResponseParameters } from "./strategyResponseParameters";
import type { StrategyResponseStrategyType } from "./strategyResponseStrategyType";
import type { StrategyResponseValidationError } from "./strategyResponseValidationError";

/**
 * Schema for strategy response.
 */
export interface StrategyResponse {
  code: string;
  code_hash: string;
  created_at: string;
  description: StrategyResponseDescription;
  execution_mode: string;
  id: string;
  indicators_used: string[];
  is_active: boolean;
  is_public: boolean;
  is_validated: boolean;
  name: string;
  parameters: StrategyResponseParameters;
  strategy_type: StrategyResponseStrategyType;
  updated_at: string;
  validation_error: StrategyResponseValidationError;
  version: number;
}